# compute_uv for SVD