--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -2764,6 +2764,10 @@
       if_true: S
       if_false: A
       default: S
+    - arg: bool compute_uv
+      if_true: S
+      if_false: N
+      default: S
 ]]
 [[
   name: _getri