--- a/aten/src/THC/generic/THCTensorMathMagma.cu
+++ b/aten/src/THC/generic/THCTensorMathMagma.cu
@@ -321,23 +321,26 @@ void THCTensor_(geev)(THCState *state, THCTensor *re_, THCTensor *rv_, THCTensor
 #endif
 }
 
-void THCTensor_(gesdd)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTensor *rv_, THCTensor *a, const char *jobu)
+void THCTensor_(gesdd)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTensor *rv_, THCTensor *a,
+                       const char *some, const char* compute_uv)
 {
 #ifdef USE_MAGMA
   THCTensor *ra_ = THCTensor_(new)(state);
-  THCTensor_(gesdd2)(state, ru_, rs_, rv_,  ra_, a, jobu);
+  THCTensor_(gesdd2)(state, ru_, rs_, rv_,  ra_, a, some, compute_uv);
   THCTensor_(free)(state, ra_);
 #else
   THError(NoMagma(gesdd));
 #endif
 }
 
-void THCTensor_(gesdd2)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTensor *rv_, THCTensor *ra_, THCTensor *a, const char *jobus)
+void THCTensor_(gesdd2)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTensor *rv_, THCTensor *ra_, THCTensor *a,
+                        const char *some, const char* compute_uv)
 {
 #ifdef USE_MAGMA
   THArgCheck(!a->is_empty() && a->dim() == 2, 2, "A should be non-empty 2 dimensional");
 
-  magma_vec_t jobz = jobus[0] == 'A' ? MagmaAllVec : jobus[0] == 'S' ? MagmaSomeVec : jobus[0] == 'O' ? MagmaOverwriteVec : MagmaNoVec;
+  char jobus = compute_uv[0] == 'N' ? 'N' : some[0];
+  magma_vec_t jobz = jobus == 'A' ? MagmaAllVec : jobus == 'S' ? MagmaSomeVec : jobus == 'O' ? MagmaOverwriteVec : MagmaNoVec;
 
   int iunused[1];
   int64_t m = a->size(0);
@@ -350,8 +353,12 @@ void THCTensor_(gesdd2)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTens
   THCTensor_(copyTensor2d)(state, a_data, a);
 
   scalar_t *rs_data = th_magma_malloc_pinned<scalar_t>(k);
-  scalar_t *ru_data = th_magma_malloc_pinned<scalar_t>(m * j);
-  scalar_t *rv_data = th_magma_malloc_pinned<scalar_t>(n * n);
+  scalar_t *ru_data = NULL;
+  scalar_t *rv_data = NULL;
+  if (jobz != MagmaNoVec) {
+    scalar_t *ru_data = th_magma_malloc_pinned<scalar_t>(m * j);
+    scalar_t *rv_data = th_magma_malloc_pinned<scalar_t>(n * n);
+  }
 
   scalar_t wkopt;
   int info;
@@ -377,18 +384,25 @@ void THCTensor_(gesdd2)(THCState *state, THCTensor *ru_, THCTensor *rs_, THCTens
   else if (info < 0)
     THError("MAGMA gesdd : Argument %d : illegal value", -info);
 
-  THCTensor_(copyArray2d)(state, rv_, rv_data, n, n);
-  THCTensor_(transpose)(state, rv_, NULL, 0, 1);
-  if (jobz != MagmaAllVec)
-    THCTensor_(narrow)(state, rv_, rv_, 1, 0, jv);
-  THCTensor_(copyArray2d)(state, ru_, ru_data, m, j);
   THCTensor_(copyArray1d)(state, rs_, rs_data, k);
   THCTensor_(copyArray2d)(state, ra_, a_data,  m, n);
+  if (jobz != MagmaNoVec) {
+    THCTensor_(copyArray2d)(state, rv_, rv_data, n, n);
+    THCTensor_(transpose)(state, rv_, NULL, 0, 1);
+    if (jobz != MagmaAllVec)
+      THCTensor_(narrow)(state, rv_, rv_, 1, 0, jv);
+    THCTensor_(copyArray2d)(state, ru_, ru_data, m, j);
+    magma_free_pinned(rv_data);
+    magma_free_pinned(ru_data);
+  } else {
+    THCTensor_(resize2d)(state, rv_, n, n);
+    THCTensor_(zero)(state, rv_);
+    THCTensor_(resize2d)(state, ru_, m, m);
+    THCTensor_(zero)(state, ru_);
+  }
 
   magma_free_pinned(work_data);
   magma_free_pinned(iwork);
-  magma_free_pinned(rv_data);
-  magma_free_pinned(ru_data);
   magma_free_pinned(rs_data);
   magma_free_pinned(a_data);
 #else