--- a/test/test_autograd.py
+++ b/test/test_autograd.py
@@ -2558,7 +2558,7 @@ class TestAutograd(TestCase):
     def test_svd_no_singularvectors(self):
         A = torch.randn(2, 2, dtype=torch.float32, requires_grad=True)
         u, s, v = torch.svd(A, compute_uv=False)
-        with self.assertRaisesRegex(RuntimeError, 'without computing the singular matrices'):
+        with self.assertRaisesRegex(RuntimeError, 'without the singular matrices'):
             torch.autograd.backward([u, s, v], [torch.ones_like(u), torch.ones_like(s), torch.ones_like(v)])
 
     def test_no_grad_copy(self):