--- a/torch/_torch_docs.py
+++ b/torch/_torch_docs.py
@@ -4412,7 +4412,7 @@ Example::
 
 add_docstr(torch.svd,
            r"""
-svd(input, some=True, out=None) -> (Tensor, Tensor, Tensor)
+svd(input, some=True, compute_uv=True, out=None) -> (Tensor, Tensor, Tensor)
 
 `U, S, V = torch.svd(A)` returns the singular value decomposition of a
 real matrix `A` of size `(n x m)` such that :math:`A = USV^T`.
@@ -4426,6 +4426,9 @@ of size :math:`\min(n, m)` containing the non-negative diagonal entries.
 
 If :attr:`some` is ``True`` (default), the returned `U` and `V` matrices will
 contain only :math:`min(n, m)` orthonormal columns.
+
+If :attr:`compute_uv` is ``False``, the returned `U` and `V` matrices will be zero matrices
+of shape :math:`(n \times n)` and :math:`(m \times m)` respectively. :attr:`some` will be ignored here.
 
 .. note:: The implementation of SVD on CPU uses the LAPACK routine `?gesdd` (a divide-and-conquer
           algorithm) instead of `?gesvd` for speed. Analogously, the SVD on GPU uses the MAGMA routine
@@ -4444,6 +4447,9 @@ contain only :math:`min(n, m)` orthonormal columns.
 .. note:: When :attr:`some` = ``False``, the gradients on ``U[:, min(n, m):]``
           and ``V[:, min(n, m):]`` will be ignored in backward as those vectors
           can be arbitrary bases of the subspaces.
+
+.. note:: When :attr:`compute_uv` = ``False``, backward cannot be performed since ``U`` and ``V`` are
+          from the forward pass is required for the backward pass.
 
 Args:
     input (Tensor): the input 2-D tensor