Adds a `compute_uv` argument that defaults to `True` for optionally computing the singular vectors during SVD.


Closes https://github.com/pytorch/pytorch/issues/12420 .