@SsnL 2018-10-10
I also wonder how much faster this is. Could you do some benchmarking on CPU (MKL) & GPU (MAGMA)? Thanks!