@vishwakftw 2018-10-11
CPU benchmarks:

size | compute_uv=True (master) | compute_uv=False (this PR)
----- | ----------------------- | ------------------------
100 x 200 | 1.88 ms ± 17.9 µs per loop | 1.11 ms ± 2.57 µs per loop
200 x 400 | 8.67 ms ± 499 µs per loop | 4.65 ms ± 591 µs per loop
500 x 250 | 13.6 ms ± 661 µs per loop | 7.1 ms ± 259 µs per loop
800 x 400 | 39.7 ms ± 2.72 ms per loop | 22.8 ms ± 1.21 ms per loop

GPU benchmarks can be expected to have the same speedup, since both use the divide and conquer algorithm (`gesdd`).